requirements.txt
Flask
Flask-GraphQL
Flask-PyMongo
graphene-mongo
graphene
flask-cors